moved parse checks that are independent of MAKE to own file
