+++ wiggle_query_language/clauses/make/parse_make/parse_make_statement.py
@@ -3,6 +3,8 @@ from typing import Optional
 from models.wql import ParsedMake
 from wiggle_query_language.clauses.make.parse_make.parse_make_statement_checks import (
     check_make_clause_spelling,
+)
+from wiggle_query_language.clauses.parsing_helpers.parse_statement_checks import (
     validate_statement,
 )
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
